# This is a combination of 2 commits. # This is the 1st commit message:
adds a change to gateway.js and rectifies e2e testing

cli/lib/gateway.js has been changed to allow necessary parameters to be passed to -config/lib/network

test-quarantine has been brought back in. This directory was nearly lost. It contains files created over the last year.
These tests are separated from the Travis tests until further work is to be done on them.
But, they were originally kept in the repository, so there are being included and are being tracked.

test-e2e-local is a variant of Dipak's test-e2e but does not use jq or yq. Travis was failing on the install of jq and yq.
Just about every option for getting Travis to install those utilties was tried to no avail. So, the the e2e-local test
is a variant the is not dependent on jq or yq but has the same functionality. Travis test should not be passing.

test-e2e is not included here since it is being held back in a branch created by Dipak.

A directory containing utilities to help manage switching package.json has been removed.

# This is the commit message #2:

Update npm-shrinkwrap.json

update local for tests.